[FEAT] Added CameraParams GetParameters from device

+++ Core/Entities/Parameters/CameraParams/Models/DB/CameraParam.cs
@@ -12,8 +12,8 @@ public partial class CameraParam : BaseEntity, IBaseEntity<CameraParam, DTOCamer
 	public string TriggerActivation { get; set; } = TriggerActivations.AnyEdge;
 	public double ExposureTime { get; set; } = 30000.0;
 	public string PixelFormat { get; set; } = PixelFormats.BayerRG8;
-	public int Width { get; set; } = 2464;
-	public int Height { get; set; } = 2056;
+	public long Width { get; set; } = 2464;
+	public long Height { get; set; } = 2056;
 	public bool AcquisitionFrameRateEnable { get; set; } = false;
 	public double Gain { get; set; } = 10;
 	public double BlackLevel { get; set; } = 50;
@@ -21,6 +21,6 @@ public partial class CameraParam : BaseEntity, IBaseEntity<CameraParam, DTOCamer
 	public double BalanceRatio { get; set; } = 2.35498;
 	public bool ConvolutionMode { get; set; } = false;
 	public double AdaptiveNoiseSuppressionFactor { get; set; } = 1;
-	public int Sharpness { get; set; } = 0;
+	public long Sharpness { get; set; } = 0;
 	public double AcquisitionFrameRate { get; set; } = 23.9798;
 }
